spotbugs

+++ src/main/java/io/cryostat/recordings/RecordingHelper.java
@@ -1603,8 +1603,7 @@ public class RecordingHelper {
         public void execute(JobExecutionContext ctx) throws JobExecutionException {
             try {
                 ActiveRecording recording =
-                        ActiveRecording.find(
-                                        "id", (long) ctx.getMergedJobDataMap().get("recordingId"))
+                        ActiveRecording.find("id", ctx.getMergedJobDataMap().get("recordingId"))
                                 .singleResult();
                 recordingHelper.stopRecording(recording).await().atMost(connectionTimeout);
             } catch (Exception e) {
